requirements.txt
python-decouple>=3.1
setuptools>=38.3
requests>=2.25.1
pyzabbix>=0.8.2
